+++ src/app/components/poll-editor/questions/multiple-choice/multiple-choice.component.spec.ts
