+++ src/app/components/poll-editor/questions/questions-styles.scss
