+++ src/app/components/poll-editor/questions/single-choice/single-choice.component.html
