+++ src/app/components/poll-editor/questions/single-choice/single-choice.component.ts
@@ -1,4 +1,3 @@
-import { I } from '@angular/cdk/keycodes';
 import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { FormArray, FormBuilder } from '@angular/forms';
