+++ src/app/modules/poll-editor/poll-editor.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SingleChoiceComponent } from '../../components/questions/single-choice/single-choice.component';
-import { MultipleChoiceComponent } from '../../components/questions/multiple-choice/multiple-choice.component';
+import { SingleChoiceComponent } from '../../components/poll-editor/questions/single-choice/single-choice.component';
+import { MultipleChoiceComponent } from '../../components/poll-editor/questions/multiple-choice/multiple-choice.component';
 import { MaterialModule } from '../material/material.module';
 import { ReactiveFormsModule } from '@angular/forms';
 import { MarkdownModule } from 'ngx-markdown';
@@ -22,4 +22,4 @@ import { MarkdownModule } from 'ngx-markdown';
     MultipleChoiceComponent
   ]
 })
-export class QuestionsModule { }
+export class PollEditorModule { }
